backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0

# Phase III: AI Chatbot dependencies
anyio>=4.6.0,<5.0.0
mcp
openai==1.54.0
httpx>=0.27.0,<1.0.0
tenacity==8.2.3
alembic==1.13.1
slowapi==0.1.9

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
